+++ test/providers/google/ai.studio.test.ts
@@ -295,7 +295,7 @@ describe('AIStudioChatProvider', () => {
         expect.objectContaining({
           method: 'POST',
           body: expect.stringContaining(
-            '"contents":{"role":"user","parts":{"text":"test prompt"}}',
+            '"contents":[{"role":"user","parts":{"text":"test prompt"}}]',
           ),
         }),
         expect.any(Number),
@@ -408,7 +408,7 @@ describe('AIStudioChatProvider', () => {
             'Content-Type': 'application/json',
           }),
           body: expect.stringContaining(
-            '"contents":[{"role":"system","parts":[{"text":"system instruction"}]},{"role":"user","parts":[{"text":"user message"}]}]',
+            '"contents":[{"role":"system","parts":{"text":"system instruction"}},{"role":"user","parts":{"text":"user message"}}]',
           ),
         }),
         expect.any(Number),
@@ -780,7 +780,7 @@ describe('AIStudioChatProvider', () => {
       expect(cache.fetchWithCache).toHaveBeenCalledWith(
         'https://rendered-generativelanguage.googleapis.com/v1beta/models/gemini-pro:generateContent?key=rendered-test-key',
         {
-          body: '{"contents":{"role":"user","parts":{"text":"What is the weather in San Francisco?"}},"generationConfig":{},"toolConfig":{"functionCallingConfig":{"mode":"AUTO","allowedFunctionNames":["get_weather"]}},"tools":[{"functionDeclarations":[{"name":"get_weather","description":"Get weather information","parameters":{"type":"OBJECT","properties":{"location":{"type":"STRING","description":"City name"}},"required":["location"]}}]}]}',
+          body: '{"contents":[{"role":"user","parts":{"text":"What is the weather in San Francisco?"}}],"generationConfig":{},"toolConfig":{"functionCallingConfig":{"mode":"AUTO","allowedFunctionNames":["get_weather"]}},"tools":[{"functionDeclarations":[{"name":"get_weather","description":"Get weather information","parameters":{"type":"OBJECT","properties":{"location":{"type":"STRING","description":"City name"}},"required":["location"]}}]}]}',
           headers: { 'Content-Type': 'application/json' },
           method: 'POST',
         },
@@ -870,7 +870,7 @@ describe('AIStudioChatProvider', () => {
       expect(cache.fetchWithCache).toHaveBeenCalledWith(
         'https://rendered-generativelanguage.googleapis.com/v1beta/models/gemini-pro:generateContent?key=rendered-test-key',
         {
-          body: '{"contents":{"role":"user","parts":{"text":"What is the weather in San Francisco?"}},"generationConfig":{},"tools":[{"functionDeclarations":[{"name":"get_weather","description":"Get weather in San Francisco","parameters":{"type":"OBJECT","properties":{"location":{"type":"STRING"}}}}]}]}',
+          body: '{"contents":[{"role":"user","parts":{"text":"What is the weather in San Francisco?"}}],"generationConfig":{},"tools":[{"functionDeclarations":[{"name":"get_weather","description":"Get weather in San Francisco","parameters":{"type":"OBJECT","properties":{"location":{"type":"STRING"}}}}]}]}',
           headers: { 'Content-Type': 'application/json' },
           method: 'POST',
         },
